requirements.txt
Django>=2.0
requests>=2.21.0
psycopg2-binary